misc updates

+++ core/src/main/scala/com/acework/js/components/bootstrap/Utils.scala
@@ -227,11 +227,11 @@ object Utils {
     scalaComponent
   }
 
-  case class Offset(top: Int, left: Int)
+  case class Offset(top: Double, left: Double)
 
   def getOffset(node: TopNode) = {
     val offset = jQuery(node).offset().asInstanceOf[js.Dynamic]
-    Offset(offset.top.asInstanceOf[Int], offset.left.asInstanceOf[Int])
+    Offset(offset.top.asInstanceOf[Double], offset.left.asInstanceOf[Double])
   }
 
 }
